util,cli: Extract Decoder to separate util package
