+++ cmd/ttn-lw-cli/commands/utils.go
@@ -23,10 +23,10 @@ import (
 	"strings"
 
 	"github.com/spf13/pflag"
-	"go.thethings.network/lorawan-stack/v3/cmd/ttn-lw-cli/internal/io"
 	"go.thethings.network/lorawan-stack/v3/cmd/ttn-lw-cli/internal/util"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+	"go.thethings.network/lorawan-stack/v3/pkg/util/io"
 )
 
 func getHost(address string) string {
