+++ cmd/ttn-lw-cli/internal/io/io.go
@@ -16,8 +16,6 @@ package io
 
 import (
 	"bufio"
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
 	"os"
@@ -104,64 +102,4 @@ func BufferedPipe(r io.Reader) (*bufio.Reader, bool) {
 		}
 	}
 	return nil, false
-}
-
-// Decoder is the interface for the functionality that reads and decodes entities
-// from an io.Reader, typically os.Stdin.
-type Decoder interface {
-	Decode(data interface{}) (paths []string, err error)
-}
-
-type jsonDecoder struct {
-	rd  *bufio.Reader
-	dec *json.Decoder
-}
-
-// NewJSONDecoder returns a new Decoder on top of r, and that uses the common JSON
-// format used in The Things Stack.
-func NewJSONDecoder(r io.Reader) Decoder {
-	rd := bufio.NewReader(r)
-	return &jsonDecoder{
-		rd:  rd,
-		dec: json.NewDecoder(rd),
-	}
-}
-
-func (r *jsonDecoder) Decode(data interface{}) (paths []string, err error) {
-	t, err := r.rd.ReadByte()
-	if err != nil {
-		return nil, err
-	}
-	if t == '{' {
-		if err := r.rd.UnreadByte(); err != nil {
-			return nil, err
-		}
-	}
-	var obj json.RawMessage
-	if err = r.dec.Decode(&obj); err != nil {
-		return nil, err
-	}
-	var m map[string]interface{}
-	if err = json.Unmarshal(obj, &m); err != nil {
-		return nil, err
-	}
-	paths = fieldPaths(m, "")
-	b := bytes.NewBuffer(obj)
-	if err = jsonpb.TTN().NewDecoder(b).Decode(data); err != nil {
-		return nil, err
-	}
-	r.rd = bufio.NewReader(io.MultiReader(r.dec.Buffered(), r.rd))
-	r.dec = json.NewDecoder(r.rd)
-	return paths, nil
-}
-
-func fieldPaths(m map[string]interface{}, prefix string) (paths []string) {
-	for path, sub := range m {
-		if m, ok := sub.(map[string]interface{}); ok {
-			paths = append(paths, fieldPaths(m, prefix+path+".")...)
-		} else {
-			paths = append(paths, prefix+path)
-		}
-	}
-	return paths
 }
